requirements.txt
# Auto-generated requirements from current Python environment
# Ensure jupyter and nbconvert CLI tools are available
jupyter
nbconvert
accessible-pygments==0.0.5
alabaster==0.7.16
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
docutils==0.21.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
greenlet==3.2.3
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-book==1.0.4.post1
jupyter-cache==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
latexcodec==3.0.1
linkify-it-py==2.0.3
markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-nb==1.2.0
myst-parser==3.0.1
myst-cli
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pip==25.1.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybtex==0.25.1
pybtex-docutils==1.0.3
pydata-sphinx-theme==0.15.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==65.5.0
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==7.4.7
sphinx-book-theme==1.1.4
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx_design==0.6.1
sphinx_external_toc==1.0.1
sphinx-jupyterbook-latex==1.0.0
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.3.1
sphinx-togglebutton==0.3.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.5
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stack-data==0.6.3
tabulate==0.9.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
uc-micro-py==1.0.3
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.23.0
